requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Django==3.0.6
django-bootstrap3==12.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
pathspec==0.8.0
pycparser==2.20
pytz==2020.1
regex==2020.5.14
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
